+++ tinted-builder-rust/src/lib.rs
@@ -3,6 +3,11 @@
 mod operations {
     pub mod build;
 }
-mod utils;
+
+mod helpers;
+
+pub mod utils {
+    pub use crate::operations::build::utils::get_scheme_files;
+}
 
 pub use crate::operations::build as operation_build;