+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - BREAKING: Remove `tinted_builder` exports since they should be
   imported from `tinted_builder` crate itself
+- Export `get_scheme_files` to allow Rust users to get a
+  `&'static [SchemeFile]` from a directory
 
 ## [0.11.1] - 2024-10-02